refactor: Centralize editor display names in core (#37)
Moves EDITOR_DISPLAY_NAMES and adds getEditorDisplayName helper to packages/core/src/utils/editor.ts. Updates CLI components to use these centralized definitions, ensuring consistent editor naming across the application.
